Refactored robots model into separate sub-modules
