+++ src/compas/robots/model/__init__.py
@@ -1,0 +1,11 @@
+from .geometry import *
+from .joint import *
+from .link import *
+from .robot import *
+
+from .geometry import __all__ as a
+from .joint import __all__ as b
+from .link import __all__ as c
+from .robot import __all__ as d
+
+__all__ = a + b + c + d
